+++ lib/chart/render/decorations/selected_item_decoration.dart
@@ -65,7 +65,8 @@ class SelectedItemDecoration extends DecorationPainter {
   @override
   Size layoutSize(BoxConstraints constraints, ChartState state) {
     final _listSize = state.data.listSize;
-    final _itemWidth = constraints.deflate(state.defaultMargin).maxWidth / _listSize;
+    final _itemWidth =
+        constraints.deflate(state.defaultMargin).maxWidth / _listSize;
 
     return Size(
       _itemWidth,
@@ -75,7 +76,8 @@ class SelectedItemDecoration extends DecorationPainter {
 
   @override
   Offset applyPaintTransform(ChartState state, Size size) {
-    final _width = (size.width - state.defaultMargin.horizontal) / state.data.listSize;
+    final _width =
+        (size.width - state.defaultMargin.horizontal) / state.data.listSize;
 
     final selectedItem = this.selectedItem;
     if (selectedItem != null) {
@@ -96,7 +98,8 @@ class SelectedItemDecoration extends DecorationPainter {
     return Offset.zero;
   }
 
-  void _drawText(Canvas canvas, Size size, double totalWidth, ChartState state) {
+  void _drawText(
+      Canvas canvas, Size size, double totalWidth, ChartState state) {
     final _item = selectedItem;
     if (_item == null) {
       return;
@@ -110,8 +113,10 @@ class SelectedItemDecoration extends DecorationPainter {
       selectedStyle,
       hasMaxWidth: false,
     );
-    final _itemWidth = max(state.itemOptions.minBarWidth ?? 0.0,
-        min(state.itemOptions.maxBarWidth ?? double.infinity, size.width - state.itemOptions.padding.horizontal));
+    final _itemWidth = max(
+        state.itemOptions.minBarWidth ?? 0.0,
+        min(state.itemOptions.maxBarWidth ?? double.infinity,
+            size.width - state.itemOptions.padding.horizontal));
 
     const _size = 2.0;
     final _maxValue = state.data.maxValue - state.data.minValue;
@@ -238,7 +243,8 @@ class SelectedItemDecoration extends DecorationPainter {
 
   void _drawItem(Canvas canvas, Size size, ChartState state) {
     canvas.drawRect(
-      Rect.fromPoints(Offset(0.0, marginNeeded().top), Offset(size.width, size.height)),
+      Rect.fromPoints(
+          Offset(0.0, marginNeeded().top), Offset(size.width, size.height)),
       Paint()
         ..color = backgroundColor
         ..blendMode = BlendMode.overlay,
@@ -262,11 +268,19 @@ class SelectedItemDecoration extends DecorationPainter {
     if (endValue is SelectedItemDecoration) {
       return SelectedItemDecoration(
         animate
-            ? (lerpDouble(selectedItem?.toDouble(), endValue.selectedItem?.toDouble(), t) ?? 0).round()
+            ? (lerpDouble(selectedItem?.toDouble(),
+                        endValue.selectedItem?.toDouble(), t) ??
+                    0)
+                .round()
             : endValue.selectedItem,
-        selectedColor: Color.lerp(selectedColor, endValue.selectedColor, t) ?? endValue.selectedColor,
-        backgroundColor: Color.lerp(backgroundColor, endValue.backgroundColor, t) ?? endValue.backgroundColor,
-        selectedStyle: TextStyle.lerp(selectedStyle, endValue.selectedStyle, t) ?? endValue.selectedStyle,
+        selectedColor: Color.lerp(selectedColor, endValue.selectedColor, t) ??
+            endValue.selectedColor,
+        backgroundColor:
+            Color.lerp(backgroundColor, endValue.backgroundColor, t) ??
+                endValue.backgroundColor,
+        selectedStyle:
+            TextStyle.lerp(selectedStyle, endValue.selectedStyle, t) ??
+                endValue.selectedStyle,
         animate: endValue.animate,
         selectedArrayIndex: endValue.selectedArrayIndex,
         showOnTop: t < 0.5 ? showOnTop : endValue.showOnTop,